+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.0.10.2 (Upcoming)
 
 * Add column-based index partitioner
+* Fix CQL query handler activation at startup
 * Fix best effort mapping of collections (#246)
 
 ## 3.0.10.1 (January 10, 2017)